# bip-reference-person/src/main/resources/openapi/openapi.yml
...
openapi: 3.0.1
info:
   version: 0.0.1-SNAPSHOT
   title: BIP Reference Person Spring Boot Service v1 API Documentation
   description: A sample Reference Person API to demonstrate features in the OpenAPI 3.0 specification
   termsOfService: https://developer.va.gov/terms-of-service
   license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local server
  - url: http://bip-reference-person-master-blue-dev.dev.bip.va.gov
    description: Development server
...
paths:
...
  "/api/v1/persons/pid":
    post:
      tags:
      - reference-person
      summary: Retrieve person information by PID from Person Service .
      description: Will return a person info object based on search by PID.
      operationId: personByPid
      requestBody: 
        description: personInfoRequest
        required: true
        content:
            application/json:
              schema:
                '$ref': '#/components/schemas/PersonInfoRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                '$ref': '#/components/schemas/PersonInfoResponse'
        '400':
          description: There was an error encountered processing the Request.  Response
            will contain a  "messages" element that will provide further information
            on the error.  This request shouldn't be retried until corrected.
          content:
            application/problem+json:
              schema:
                '$ref': '#/components/schemas/ProviderResponse'
        '403':
          description: The request is not authorized.  Please verify credentials used
            in the request.
          content:
            application/problem+json:
              schema:
                '$ref': '#/components/schemas/ProviderResponse'
        '500':
          description: There was an error encountered processing the Request.  Response
            will contain a  "messages" element that will provide further information
            on the error.  Please retry.  If problem persists, please contact support
            with a copy of the Response.
          content:
            application/problem+json:
              schema:
                '$ref': '#/components/schemas/ProviderResponse'
      security:
        - bearerAuth: []
      deprecated: false
  # The token path should not require any changes
...
components:
...
  securitySchemes:
    Authorization:
      type: apiKey
      name: Authorization
      in: header
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
  schemas:
...
    PersonInfo:
      type: object
      title: PersonInfo
      description: Model for data contained in the response from the Person Service
      properties:
        fileNumber:
          type: string
          example: 912444689
          description: The persons file number
        firstName:
          type: string
          example: JANE
          description: The persons first name
        lastName:
          type: string
          example: DOE
          description: The persons last name
        middleName:
          type: string
          example: M
          description: The persons middle name
        participantId:
          type: integer
          format: int64
          example: 6666345
          minimum: 1
          description: The persons participant ID
        socSecNo:
          type: string
          example: 912444689
          description: The persons SSN
    PersonInfoRequest:
      title: PersonInfoRequest
      description: Model for data to request person information from the Person Service
      allOf:
        - '$ref': '#/components/schemas/ProviderRequest'
        - type: object
          properties:
            participantID:
              type: integer
              format: int64
              example: 6666345
              description: The Participant ID of the person for whom to retrieve data
              minimum: 1
              exclusiveMinimum: false
          required:
            - participantID
    PersonInfoResponse:
      title: PersonInfoResponse
      description: Model for the response from the Person Service
      allOf:
       - '$ref': '#/components/schemas/ProviderResponse'
       - type: object
         # all other properties specific to PersonInfoResponse
         properties:
            personInfo:
              description: The object representing the person information
              '$ref': '#/components/schemas/PersonInfo'
    # Schema objects below are from BIP Framework to be declared.
    # These objects need to be mapped in <importMappings> section
    # of openapi-generator-maven-plugin so that no code is generated 
    # for these classes
    Message:
      type: object
      title: Message
      description: Model that identifies a single individual used in the security context
      required:
      - key
      - severity
      properties:
        key:
          # possible values enumerated by gov.va.bip.framework.messages.MessageKeys
          type: string
        severity:
          # derived from gov.va.bip.framework.messages.MessageSeverity
          type: string 
          enum:
            - INFO
            - WARN
            - ERROR
        status:
          type: integer
          format: int32
        text:
          type: string
        timestamp:
          type: string
          format: date-time
          example: yyyy-MM-dd'T'HH:mm:ss.SSS
...
    ProviderResponse: 
      type: object
      properties:
        messages:
          type: array
          items:
            '$ref': '#/components/schemas/Message'
      title: ProviderResponse
    ProviderRequest: 
      type: object
      title: ProviderRequest